requirements.txt
mysql-connector-python==2.1.3
python-sql==0.8
